optional separation of functions declarations for linux interface
